+++ common/src/main/kotlin/com/cobblemon/mod/common/api/spawning/condition/SpawningCondition.kt
@@ -18,8 +18,11 @@ import com.cobblemon.mod.common.api.spawning.detail.SpawnDetail
 import com.cobblemon.mod.common.util.Merger
 import com.cobblemon.mod.common.util.math.orMax
 import com.cobblemon.mod.common.util.math.orMin
+import com.mojang.datafixers.util.Either
+import net.minecraft.registry.tag.TagKey
 import net.minecraft.util.Identifier
 import net.minecraft.world.biome.Biome
+import net.minecraft.world.gen.structure.Structure
 
 /**
  * The root of spawning conditions that can be applied to a spawning context. What type
@@ -54,6 +57,7 @@ abstract class SpawningCondition<T : SpawningContext> {
     var timeRange: TimeRange? = null
     var labels: MutableList<String>? = null
     var labelMode = ANY
+    var structures: MutableList<Either<Identifier, TagKey<Structure>>>? = null
 
     @Transient
     var appendages = mutableListOf<AppendageCondition>()
@@ -101,6 +105,16 @@ abstract class SpawningCondition<T : SpawningContext> {
             return false
         } else if (appendages.any { !it.fits(ctx, detail) }) {
             return false
+        } else if (structures != null && structures!!.isNotEmpty() &&
+            structures!!.let { structures ->
+                val structureAccess = ctx.world.structureAccessor
+                val cache = ctx.getStructureCache(ctx.position)
+                return@let structures.none {
+                    it.map({ cache.check(structureAccess, ctx.position, it) }, { cache.check(structureAccess, ctx.position, it) })
+                }
+            }
+        ) {
+            return false
         }
 
         return true
@@ -121,5 +135,6 @@ abstract class SpawningCondition<T : SpawningContext> {
         minLight = merger.mergeSingle(minLight, other.minLight)
         maxLight = merger.mergeSingle(maxLight, other.maxLight)
         timeRange = merger.mergeSingle(timeRange, other.timeRange)
+        structures = merger.merge(structures, other.structures)?.toMutableList()
     }
 }